updating .env.example and creating a template for trip errors
