+++ .env.example
@@ -6,11 +6,16 @@ DB_HOST=XXX
 DB_NAME=XXX
 
 JWT_KEY=XXX
-EMAIL=xxx
+EMAIL=info.barefoot.nomad@gmail.com
 EMAIL_PASSWORD= xxx
 
 SIGNUP_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwiZW1haWwiOiJmaXJhZHVrQGdtYWlsLmNvbSIsImlhdCI6MTU4NTQ4OTE2OH0.0Npc12n_FRWhyADnYYNIhBcZ2JlIYCyc3H5UcmHAb6w
 SIGNUP_EXPIRED_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwiZW1haWwiOiJmaXJoa2pAZ21haWwuY29tIiwiaWF0IjoxNTg1NDI4MTA5LCJleHAiOjE1ODU0NDI1MDl9.uNGkjCjid1ujW846Gcl58gntlmjgxxyvL0ikaqQgssQ
+
+PASSING_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZnVsbE5hbWUiOiJJUkFEVUtVTkRBIEZpYWNyZSIsImVtYWlsIjoiZmlyYWR1a0B5YWhvby5jb20iLCJ1c2VyUm9sZSI6InVzZXIiLCJpc1ZlcmlmaWVkIjp0cnVlLCJpYXQiOjE1ODYzNjY2MzV9.U6nQcNlqwjWojF1i6KgNZ7_yuMmNCBDQu_xCQTr6AM4
+UN_VERIFIED_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NCwiZnVsbE5hbWUiOiJOZG9saSBKYWNrIiwiZW1haWwiOiJpYmlyYXlpQGdtYWlsLmNvbSIsInVzZXJSb2xlIjoidXNlciIsImlzVmVyaWZpZWQiOnRydWUsImlhdCI6MTU4NjQyMjA0MX0.O6pC79-Fp0b-ZwT7HTzO8_BL_JhH6ZQ9PHJJHGSK_iE
+NOTFOUND_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwiZW1haWwiOiJyYWR1a0BnbWFpbC5jb20iLCJpYXQiOjE1ODYyMDM1MTR9.H6WyuW03GjPEWC3hqzc9ThF0JF1pXksNsoFyo9hUoK0
+NOTFOUND2_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
 
 VERIFY_URL=http://localhost:5000/api/v1/users/verify
 URL_DEV=https://warriorz.herokuapp.com
