+++ logs/all.log.14-04-2020
@@ -32,3 +32,24 @@
 {"service":"user-service","error":"Incorrect use of special characters","tip":"Please avoid using numbers or special characters characters that looks like = or /","path":"travelReason","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
 {"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::ffff:127.0.0.1"}
 {"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::1"}
+{"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::1"}
+{"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::1"}
+{"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 401, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 401, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 401, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"debug","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 401, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"info","message":"METHOD: GET, and the URL: /api/v1/trips/cities, STATUSCODE: 200 IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","error":"Origin City is not supported by Barefoot Nomad","tip":"Choose from the cities we have available....","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","error":"Destination City is not supported by Barefoot Nomad","tip":"Choose from the cities we have available....","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","error":"Incorrect Date Format","format":"YYYY-MM-DD","path":"date","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","error":"Incorrect use of special characters","tip":"Please avoid using numbers or special characters characters that looks like = or /","path":"travelReason","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"info","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 200 IP-address: ::ffff:127.0.0.1"}
+{"service":"user-service","level":"error","message":"METHOD: POST, and the URL: /api/v1/trips, STATUSCODE: 400, IP-address: ::ffff:127.0.0.1"}
